// src/types/renshuuApiTypes.ts
export interface paths {
// ...
    "/grammar/search": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Search renshuu's grammar dictionary
         * @description Returns matching grammar.
         */
        get: {
            parameters: {
                query: {
                    /** @description The search value (Japanese or English) */
                    value: string;
                    /** @description Page of results (count starts from 1) */
                    pg?: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example 1 */
                            result_count?: number;
                            /** @example 2 */
                            total_pg?: number;
                            /** @example 20 */
                            per_pg?: number;
                            /** @example 1 */
                            pg?: number;
                            grammar?: components["schemas"]["Grammar"][];
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/grammar/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Lookup grammar
         * @description Returns specific grammar.
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    /** @description The grammar Id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Grammar"];
                    };
                };
            };
        };
        /** Add grammar to schedule or list */
        put: {
            parameters: {
                query?: {
                    /** @description The ID of the list to add to */
                    list_id?: string;
                    /** @description The ID of the schedule to add to */
                    sched_id?: string;
                };
                header?: never;
                path: {
                    /** @description The grammar Id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully added. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description List or schedule could not be found. */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description List or schedule already contained this grammar. */
                409: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        post?: never;
        /** Remove grammar from schedule or list */
        delete: {
            parameters: {
                query?: {
                    /** @description The ID of the list to remove from */
                    list_id?: string;
                    /** @description The ID of the schedule to remove from */
                    sched_id?: string;
                };
                header?: never;
                path: {
                    /** @description The grammar Id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully removed. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description List or schedule could not be found. */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description List or schedule did not contain this grammar. */
                409: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}